requirements.txt
python-kasa==0.7.1
viam-sdk==0.27.0
